+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Add `Patron comments` field to request. Refs UIREQ-530.
 * Fix fulfillment misspelling. Fixes UIREQ-567.
 * Add `Patron comments` field to search results CSV export. Refs UIREQ-549.
+* Show `lastCheckedInDateTime` token for staff slips in locale format and date/time. Refs UIREQ-495.
 
 ## [4.0.1](https://github.com/folio-org/ui-requests/tree/v4.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v4.0.0...v4.0.1)